src/requirements.txt
boto3
opencv-python
facenet-pytorch
Pillow